Cast HDU data to int32 before hashing so ZERO_HASH matches
``ZERO_HASH`` is precomputed from ``np.zeros((100, 100), dtype=np.int32)``,
but ``_hashFile`` hashed whatever dtype the FITS HDU stored (typically
``uint16``, ``float32``, or ``int32`` depending on the instrument).
The byte representation of an all-zero array differs between dtypes,
so the zero-file detection only worked for instruments that happened
to store ``int32`` pixel data and silently failed for all others.
Cast the sliced data to ``int32`` before hashing so any genuinely
all-zero region produces ``ZERO_HASH`` regardless of the HDU's dtype.

Written by Claude Code 4.6 Opus (high-effort).

+++ python/lsst/summit/extras/headerFunctions.py
@@ -154,7 +154,9 @@ def _hashFile(fileToHash: Any, dataHdu: int | str, sliceToUse: slice) -> str:
     hashStr : `str`
         Hex-encoded SHA-256 digest of the sliced data.
     """
-    data = fileToHash[dataHdu].data[sliceToUse, sliceToUse].tobytes()
+    # Cast to a fixed dtype so an all-zero array hashes to ZERO_HASH
+    # regardless of the HDU's native pixel dtype.
+    data = fileToHash[dataHdu].data[sliceToUse, sliceToUse].astype(np.int32).tobytes()
     h = _hashData(data)
     return h
